pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
# Core test runner
pytest = "^8.3.5"
pytest-flask = "^1.3.0"

# Coverage support
[tool.poetry.group.testing.dependencies]
coverage = "^7.10.1"
pytest-cov = "^6.2.1"

[project]
name             = "event-finder"
version          = "0.0.1"
description      = "A Flask-based Event Finder API with PGVector-powered RAG"
readme           = "README.md"
license          = { text = "None" }
authors          = [{ name="Mile", email="demo@xml.com" }]
requires-python  = ">=3.12,<4.0"

dependencies = [
  "alembic==1.16.1",
  "aniso8601==10.0.1",
  "attrs==25.3.0",
  "bcrypt==4.3.0",
  "blinker==1.9.0",
  "click==8.1.8",
  "colorama==0.4.6",
  "dependency-injector==4.46.0",
  "Flask==3.1.0",
  "Flask-Bcrypt==1.0.1",
  "Flask-HTTPAuth==4.8.0",
  "Flask-JWT-Extended==4.7.1",
  "flask-marshmallow==1.3.0",
  "flask-restx==1.3.0",
  "Flask-SQLAlchemy==3.1.1",
  "greenlet==3.2.1",
  "gunicorn==23.0.0",
  "importlib_resources==6.5.2",
  "iniconfig==2.1.0",
  "itsdangerous==2.2.0",
  "Jinja2==3.1.6",
  "jsonschema==4.23.0",
  "jsonschema-specifications==2025.4.1",
  "MarkupSafe==3.0.2",
  "marshmallow==4.0.0",
  "marshmallow-sqlalchemy==1.4.2",
  "packaging==25.0",
  "pluggy==1.6.0",
  "PyJWT==2.10.1",
  "python-dotenv==1.1.0",
  "pytz==2025.2",
  "referencing==0.36.2",
  "rpds-py==0.24.0",
  "SQLAlchemy==2.0.40",
  "typing_extensions==4.13.2",
  "Werkzeug==3.1.3",

  # ─── Migration tooling ─────────────────────────────────────────────────────
  "Flask-Migrate==4.1.0",
  "Mako==1.3.10",

  # ─── extras for RAG & PostgresVector ──────────────────────────────────────
  "ollama==0.4.9",
  "openai>=1.0.0,<2.0.0",
  "psycopg2-binary",
  "pgvector",

  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "build==1.2.2.post1",
  "CacheControl==0.14.3",
  "certifi==2025.4.26",
  "charset-normalizer==3.4.2",
  "cleo==2.1.0",
  "crashtest==0.4.1",
  "distlib==0.3.9",
  "distro==1.9.0",
  "dulwich==0.22.8",
  "fastjsonschema==2.21.1",
  "filelock==3.18.0",
  "findpython==0.6.3",
  "fsspec==2025.5.1",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "huggingface-hub==0.32.1",
  "idna==3.10",
  "installer==0.7.0",
  "jaraco.classes==3.4.0",
  "jaraco.context==6.0.1",
  "jaraco.functools==4.1.0",
  "jiter==0.10.0",
  "joblib==1.5.1",
  "keyring==25.6.0",
  "more-itertools==10.7.0",
  "mpmath==1.3.0",
  "msgpack==1.1.0",
  "networkx==3.4.2",
  "numpy==2.3.0",
  "pbs-installer==2025.5.17",
  "pillow==11.2.1",
  "pkginfo==1.12.1.2",
  "platformdirs==4.3.8",
  "pydantic==2.11.4",
  "pydantic_core==2.33.2",
  "pyproject_hooks==1.2.0",
  "pywin32-ctypes==0.2.3",
  "PyYAML==6.0.2",
  "RapidFuzz==3.13.0",
  "regex==2024.11.6",
  "requests==2.32.3",
  "requests-toolbelt==1.0.0",
  "safetensors==0.5.3",
  "scikit-learn==1.6.1",
  "scipy==1.15.3",
  "sentence-transformers==4.1.0",
  "setuptools==80.8.0",
  "shellingham==1.5.4",
  "sniffio==1.3.1",
  "sympy==1.14.0",
  "threadpoolctl==3.6.0",
  "tokenizers==0.21.1",
  "tomlkit==0.13.2",
  "torch==2.7.0",
  "tqdm==4.67.1",
  "transformers==4.52.3",
  "trove-classifiers==2025.5.9.12",
  "typing-inspection==0.4.1",
  "urllib3==2.4.0",
  "virtualenv==20.31.2",
  "wheel==0.45.1",
  "zstandard==0.23.0",
]

[tool.poetry.extras]
testing = ["coverage", "pytest-cov"]
